♻️ Extract actions (#779)
